requirements.txt
asgiref==3.8.1
certifi==2025.4.26
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
PyJWT==2.9.0
sqlparse==0.5.3
